Refactor: mode dto to model package (#70)
