+++ internal/model/kserve.go
@@ -1,4 +1,4 @@
-package dto
+package model
 
 // InferenceReq는 kserve에 요청을 보내기 위한 구조체입니다.
 type InferenceReq struct {
